# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wasabi Card API",
        "description": "Backend integration layer for the Wasabi Card Open API platform.\n\n## Authentication\nAll endpoints require an **X-API-KEY** header.\n\n## Response Envelope\nEvery response uses: `{ success, code, msg, data }`\n\n## Rate Limiting\n60 requests per minute per API key.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current environment"
        }
    ],
    "paths": {
        "/api/v1/accounts/assets": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Assets",
                "description": "Returns all accounts with live balances for the merchant. Data is real-time and never cached. Source: Wasabi Card /merchant/core/mcb/account/info",
                "operationId": "getAssets",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AccountObject"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Upstream Wasabi API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Account"
                ],
                "summary": "Account List",
                "description": "Returns a paginated list of accounts with optional filters. Data is real-time and never cached. Source: Wasabi Card /merchant/core/mcb/account/list",
                "operationId": "getAccountList",
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "query",
                        "description": "Filter by account ID",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 19847563867367666
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by account type: WALLET (Wallet Account) or MARGIN (Margin Account)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "WALLET",
                            "enum": [
                                "WALLET",
                                "MARGIN"
                            ]
                        }
                    },
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "Current page number. Default is 1",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of records per page. Default 10, maximum 10",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10,
                            "maximum": 10,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "msg": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "data": {
                                            "properties": {
                                                "total": {
                                                    "description": "Total matching records",
                                                    "type": "integer",
                                                    "example": 21
                                                },
                                                "records": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/AccountObject"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Upstream Wasabi API error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountObject": {
                "description": "A single merchant account with live balance information",
                "properties": {
                    "accountId": {
                        "description": "Unique account identifier",
                        "type": "string",
                        "example": "19847563867367666"
                    },
                    "accountName": {
                        "description": "Account display name",
                        "type": "string",
                        "example": "wallet9023"
                    },
                    "accountType": {
                        "description": "Account type (e.g. WALLET, MARGIN)",
                        "type": "string",
                        "example": "WALLET"
                    },
                    "currency": {
                        "description": "ISO 4217 currency code",
                        "type": "string",
                        "example": "USD"
                    },
                    "totalBalance": {
                        "description": "Total balance including frozen funds",
                        "type": "number",
                        "example": 100
                    },
                    "availableBalance": {
                        "description": "Spendable balance",
                        "type": "number",
                        "example": 100
                    },
                    "frozenBalance": {
                        "description": "Frozen / reserved balance",
                        "type": "number",
                        "example": 0
                    },
                    "digital": {
                        "description": "Number of decimal places for the currency",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "code": {
                        "type": "integer",
                        "example": 401
                    },
                    "msg": {
                        "type": "string",
                        "example": "Invalid or missing API key."
                    },
                    "data": {
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "API key issued to third-party developers.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}